Rename parser dir to ast
