+++ Makefile
@@ -30,7 +30,6 @@ minor: ## Publish new minor version
 
 minor-core: ## Publish new efx-core minor version
 	@cargo test -p efx-core
-	@make check
 	@cargo fmt --all
 	@cargo release minor -p efx-core
 	@cargo publish -p efx-core
