+++ efx-core/src/lib.rs
@@ -1,4 +1,4 @@
+pub mod ast;
 pub mod attr;
-pub mod parser;
 
-pub use parser::{Attr, Element, Interpolation, Node, ParseError, Parser, Text, parse_str};
+pub use ast::{Attr, Element, Interpolation, Node, ParseError, Parser, Text, parse_str};
